Refactored global constants into webelements.py for autocompletion, moved non-main functions into application package
